third_party/WebKit/LayoutTests/inspector/sources/debugger/internal-script-flag-for-injected-script-expected.txt

CONSOLE MESSAGE: line 2: 42
Tests that injected script has isInternalScript flag.

Internal script
Not internal script
